+++ tests/test_group_data.py
@@ -52,9 +52,12 @@ def test_group_definitions():
         assert group.id[-1] in ascii_lowercase + digits
         assert set(group.id).issubset(ascii_lowercase + digits + "_")
         assert group.id.islower()
-        # Only the group referencing all platforms is allowed to starts with "all_"
-        # prefix.
-        assert group.id == "all_platforms" or not group.id.startswith("all_")
+        # Only the group referencing all platforms and its derivate are allowed to
+        # starts with "all_" prefix.
+        assert group.id in (
+            "all_platforms",
+            "all_platforms_without_ci",
+        ) or not group.id.startswith("all_")
         assert group.id not in ALL_PLATFORM_IDS
         assert group.id in ALL_GROUP_IDS
         assert group.id in ALL_IDS